+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: lyaiche <lyaiche@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/05 14:12:16 by lyaiche           #+#    #+#             */
-/*   Updated: 2021/10/05 14:24:42 by lyaiche          ###   ########.fr       */
+/*   Updated: 2021/10/14 16:37:23 by lyaiche          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*memcpy(void *dest, const void *src, size_t n)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (n > count)
