+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: lyaiche <lyaiche@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 18:56:43 by lucasyaiche       #+#    #+#             */
-/*   Updated: 2021/10/05 14:02:57 by lyaiche          ###   ########.fr       */
+/*   Updated: 2021/10/14 16:37:29 by lyaiche          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
+#include "libft.h"
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (n > count)
